Extract WebSocket transport; move HTTP viewer into rumoca-viz-web
- new crate rumoca-transport-websocket: broadcast WS server
  (run_broadcast_server) with the same behavior as the inlined code —
  single connection, latest-only state push, client→server realtime
  toggle command
- rumoca-viz-web gains a sim_viewer_server module owning the HTTP
  server + bundled viewer.html + a minimal placeholder scene (used when
  no --scene is passed). tiny_http moves from sim-fb to viz-web.
- sim-fb: deletes server.rs, moves viewer.html to viz-web/web/,
  default_scene.js (quadrotor-specific) to examples/quadrotor_sil/
  so the quadrotor example is complete like rover. Drops tungstenite,
  tiny_http from direct deps.

All 35 sim-fb tests pass; workspace clippy --all-features clean.
