+++ crates/rumoca-sim-fb/Cargo.toml
@@ -13,14 +13,14 @@ rumoca-codec-flatbuffers = { workspace = true }
 rumoca-session = { workspace = true }
 rumoca-solver-diffsol = { workspace = true }
 rumoca-transport-udp = { workspace = true }
+rumoca-transport-websocket = { workspace = true }
+rumoca-viz-web = { workspace = true }
 
 # External dependencies
 anyhow = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 toml = { workspace = true }
-tungstenite = "0.26"
-tiny_http = "0.12"
 signal-hook = "0.3"
 crossterm = "0.28"
 gilrs = "0.11"
