+++ crates/rumoca-sim-fb/src/lib.rs
@@ -6,7 +6,6 @@
 
 pub mod config;
 pub mod input_engine;
-pub mod server;
 pub mod signal_mapper;
 pub mod sim_loop;
 
@@ -46,7 +45,7 @@ pub fn run(args: SimFbArgs) -> Result<()> {
     if args.scene_script.is_some() {
         eprintln!("  Scene: custom");
     } else {
-        eprintln!("  Scene: default (quadrotor example)");
+        eprintln!("  Scene: placeholder (pass --scene to render a vehicle)");
     }
 
     // Load FlatBuffer schemas (only when configured for autopilot coupling).
@@ -87,16 +86,19 @@ pub fn run(args: SimFbArgs) -> Result<()> {
     .context("Failed to create simulation stepper")?;
     eprintln!("  Inputs: {:?}", stepper.input_names());
 
-    // Start HTTP server in background
+    // Start HTTP viewer server in background
     let http_port = args.http_port;
     let ws_port = args.ws_port;
     let scene_script = args.scene_script.clone();
     let debug = args.debug;
     thread::spawn(move || {
-        if let Err(e) =
-            server::start_http_server(http_port, ws_port, scene_script.as_deref(), debug)
-        {
-            eprintln!("HTTP server error: {}", e);
+        if let Err(e) = rumoca_viz_web::start_viewer_server(
+            http_port,
+            ws_port,
+            scene_script.as_deref(),
+            debug,
+        ) {
+            eprintln!("HTTP server error: {e}");
         }
     });
