+++ crates/rumoca-viz-web/Cargo.toml
@@ -7,7 +7,9 @@ license.workspace = true
 repository.workspace = true
 
 [dependencies]
+anyhow = { workspace = true }
 serde_json = { workspace = true, features = ["unbounded_depth"] }
+tiny_http = "0.12"
 
 [lints]
 workspace = true
