+++ examples/quadrotor_sil/quadrotor_scene.js
